requirements.txt
# Web framework
fastapi==0.117.1

# ASGI server (includes uvloop, httptools, websockets, etc.)
uvicorn[standard]==0.36.1

# AWS SDK
boto3==1.40.36

# Templates
jinja2==3.1.4

# Imaging
Pillow==11.3.0

# Plotting
matplotlib==3.9.3

# HTTP client (needed by routes.py)
httpx==0.27.2

# multipart parser for form/file uploads
python-multipart>=0.0.7,<0.1.0